restructure controller into multiple files
