+++ backend/src/controllers/mod.rs
@@ -1,0 +1,4 @@
+pub mod recipe_controller;
+pub mod user_controller;
+
+pub use self::{recipe_controller::*, user_controller::*};
